Louis/parallel-nodes-fixes (#33)
* Fix parallel nodes execution

This fix comes with a limitation: only 1 `shell.execute()` command can
be present in the body of a node which runs in parallel with other
nodes.

* Fix Python path for Windows

Fix #28.

* Fix Python shell execution for nodes

This is only for nodes concerning random number generation - will fix
for other nodes in future commit.

* Add newline to end of snippets

This is necessary so that newlines are automatically added when scripts
are appended to one another.

* Move validateInput function up

* Minor fixes/changes

* Printing the shell from simulator node

* Fix PythonShell execution for nodes

Using the current fix, only one execution statement can be present at
each node.

Co-authored-by: Theo Reignier <ucabteb@ucl.ac.uk>