Add workaround for block command execution error
Sometimes when executing commands contained in structures (e.g. loops,
conditionals, etc) the output will be a single '.' retrieved from
stderr. This behaviour is inconsistent and it's unclear what conditions
causes it to occur, so this commit acts as a workaround for the problem
until a more robust fix can be found.