Louis/bug-fixes (#60)
* Fix quantum register sending on error

The quantum register was sending the output to the next node even if an
error occurred when executing the Python script.

* Fix #55

Fix all incorrectly formatted strings.

* Fix #59

Ensure that qubits are reset regardless of whether execution succeeded
or failed.

* Add reset functions

* Revert "Fix quantum register sending on error"

This reverts commit d195f3387821ef414e6c8e7a5154e07177b6aee8.
For some reason this doesn't work and I don't know why.

* Disable flaky tests

Disabling these tests until issue #58 is resolved.

* Return PythonShell error as an Error object

* Revert "Return PythonShell error as an Error object"

This reverts commit 8f2095789730db4243a5a394526cb8d65364d1ba.
This made it so the Python welcome message began with 'Error:'. Will
deal with this at a later date.