Louis/command-tests (#88)
* Add commandExecuted test util function

Function for testing whether a command has been executed by a node.

* Add execute script tests for measure and simulator

* Add inputId() and outputId() methods

These methods can be used to easily retrieve the IDs of the first and
last nodes. Updated the commandExecuted() function to reflect this
change.

* Fix multiple nodes of the same type being added

* Remove assumed wires array

Remove wires being put into an array in add() method. This causes
complications if a node is connected to multiple other nodes.

* Fix PythonShell tests not starting fresh

* Add execute command tests

Includes:
- Bloch sphere
- Circuit diagram
- Classical register
- Hadamard gate
- Quantum circuit
- Quantum register
- Toffoli gate

* Run Node.js CI workflow on test/ file

* Revert "Fix PythonShell tests not starting fresh"

This reverts commit 8f560f9d94222042ca6ce4fd2a2ce954795450fe.

* Increase timeout limit

* Add before hook to PythonShell constructor tests

Stops the shell before running tests.

* Fix done() being called multiple times

* Add Toffoli gate command executed test

* Add multi-controlled-u-gate test suite

* Fix multiple done calls on error

* Add ibm-quantum-system execute command test

This command is disabled by default as it requires an API key to run.
For this to run on the CI, GitHub secrets could be used. To run this
locally, an API key should be assigned to the API_TOKEN constant.

* Add more execute command tests

Add tests for:
- barrier_spec.js
- cnot-gate_spec.js
- controlled-u-gate_spec.js
- identity-gate_spec.js
- multi-controlled-u-gate_spec.js
- not-gate_spec.js
- phase-gate_spec.js
- reset_spec.js
- rotation-gate_spec.js
- swap_spec.js
- unitary-gate_spec.js